# apps/backend/src/config/swagger.yaml
openapi: 3.0.3
info:
  title: Plex Exporter API
  description: |
    Backend API for Plex Exporter application providing access to media metadata, statistics, and hero content.

    ## Features
    - Media browsing (movies and TV series)
    - Advanced filtering and search
    - Statistics and analytics
    - Hero content pipeline with TMDB enrichment
    - Rate limiting and caching for performance

    ## Rate Limiting
    - General endpoints: 100 requests per 15 minutes
    - Search endpoints: 30 requests per 1 minute
    - Hero pipeline: 10 requests per 5 minutes

    ## Caching
    - Stats/Recent: 1 minute cache
    - Lists/Search: 5 minutes cache
    - Details: 15 minutes cache

    Check `X-Cache` header to see if response was cached (HIT/MISS).
  version: 1.0.0
  contact:
    name: Plex Exporter
    url: https://github.com/schinkenmeier/plex-exporter
servers:
  - url: http://localhost:4001
    description: Local development server
...
components:
  schemas:
...
    MediaExtended:
      allOf:
        - $ref: '#/components/schemas/MediaBase'
        - type: object
          properties:
            genres:
              type: array
              items:
                type: string
              nullable: true
              example: ["Drama", "Kriegsfilm", "Action"]
            directors:
              type: array
              items:
                type: string
              nullable: true
              example: ["Sam Mendes"]
            countries:
              type: array
              items:
                type: string
              nullable: true
              example: ["United Kingdom", "United States of America"]
            collections:
              type: array
              items:
                type: string
              nullable: true
            rating:
              type: number
              format: float
              nullable: true
            audienceRating:
              type: number
              format: float
              nullable: true
            contentRating:
              type: string
              nullable: true
              example: "de/16"
            studio:
              type: string
              nullable: true
            tagline:
              type: string
              nullable: true
            duration:
              type: integer
              nullable: true
              description: Duration in milliseconds
            originallyAvailableAt:
              type: string
              format: date
              nullable: true
    Stats:
      type: object
      properties:
        totalMovies:
          type: integer
          description: Total number of movies
        totalSeries:
          type: integer
          description: Total number of TV series
        totalItems:
          type: integer
          description: Total number of media items
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
...
  responses:
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  message:
                    type: string
                    example: "You have exceeded the rate limit. Please try again later."
                  statusCode:
                    type: integer
                    example: 429
                  details:
                    type: object
                    properties:
                      type:
                        type: string
                        example: "Too Many Requests"
                      retryAfter:
                        type: string
                        description: Time until rate limit resets
              meta:
                type: object
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    MediaType:
      name: type
      in: query
      description: Filter by media type
      schema:
        type: string
        enum: [movie, tv]
    Year:
      name: year
      in: query
      description: Filter by specific year
      schema:
        type: integer
        minimum: 1800
        maximum: 2100
    YearFrom:
      name: yearFrom
      in: query
      description: Filter by year range (from)
      schema:
        type: integer
        minimum: 1800
        maximum: 2100
    YearTo:
      name: yearTo
      in: query
      description: Filter by year range (to)
      schema:
        type: integer
        minimum: 1800
        maximum: 2100
    Limit:
      name: limit
      in: query
      description: Maximum number of results
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    SortBy:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        enum: [title, year, added, updated]
        default: title
    SortOrder:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum: [asc, desc]
        default: asc
paths:
...
  /api/v1/stats:
    get:
      summary: Get statistics
      description: Get overall statistics about media library
      tags:
        - Stats
      responses:
        '200':
          description: Statistics retrieved successfully
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
            RateLimit-Limit:
              description: Request limit per window
              schema:
                type: integer
            RateLimit-Remaining:
              description: Requests remaining in window
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stats'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/movies:
    get:
      summary: List all movies
      description: Get a list of all movies with extended metadata
      tags:
        - Media
      responses:
        '200':
          description: Movies retrieved successfully
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaExtended'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/movies/{id}:
    get:
      summary: Get movie by ID
      description: Get detailed information about a specific movie
      tags:
        - Media
      parameters:
        - name: id
          in: path
          required: true
          description: Plex rating key
          schema:
            type: string
      responses:
        '200':
          description: Movie retrieved successfully
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaExtended'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/series:
    get:
      summary: List all series
      description: Get a list of all TV series with extended metadata
      tags:
        - Media
      responses:
        '200':
          description: Series retrieved successfully
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaExtended'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/series/{id}:
    get:
      summary: Get series by ID
      description: Get detailed information about a specific TV series
      tags:
        - Media
      parameters:
        - name: id
          in: path
          required: true
          description: Plex rating key
          schema:
            type: string
      responses:
        '200':
          description: Series retrieved successfully
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaExtended'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/filter:
    get:
      summary: Filter media
      description: Filter and sort media with various criteria
      tags:
        - Search
      parameters:
        - $ref: '#/components/parameters/MediaType'
        - $ref: '#/components/parameters/Year'
        - $ref: '#/components/parameters/YearFrom'
        - $ref: '#/components/parameters/YearTo'
        - name: search
          in: query
          description: Text search in title/summary
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
      responses:
        '200':
          description: Filtered results
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaExtended'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/search:
    get:
      summary: Search media
      description: Search media by title or summary text
      tags:
        - Search
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - $ref: '#/components/parameters/MediaType'
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: Search results
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/recent:
    get:
      summary: Get recent media
      description: Get recently added media items
      tags:
        - Media
      parameters:
        - name: limit
          in: query
          description: Maximum number of results
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
        - $ref: '#/components/parameters/MediaType'
      responses:
        '200':
          description: Recent media retrieved
          headers:
            X-Cache:
              description: Cache status (HIT or MISS)
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/MediaExtended'
                  count:
                    type: integer
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...